backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-generativeai==0.8.5
asyncpg==0.29.0
pydantic==2.11.7
python-dotenv==1.0.1 